+++ sentry-opentelemetry/lib/sentry-opentelemetry.rb
@@ -1,0 +1,6 @@
+require "sentry-ruby"
+require "opentelemetry-sdk"
+
+require "sentry/opentelemetry/version"
+require "sentry/opentelemetry/span_processor"
+require "sentry/opentelemetry/propagator"